[IMP] packagemore_ecommerce_pricelists: Set minimum qty and show price in website product list
When a product has more than one pricelist line, and all minimum quantities are
greater than 1, the website product list view would still show "Please contact
us for quote", which was not expected.
This commit takes the smallest minimum quantity defined for a product in a
pricelist, and displays the corresponding price in the website product list view.

Additionally, when you enter the form view of a product in the website, the
smallest minimum quantity in the pricelist is automatically set.

This is the commit message #2:

[IMP] packagemore_ecommerce_pricelists: Refactor code after 1st feedback on PR

This is the commit message #3:

[IMP] packagemore_ecommerce_pricelists: Refactor code after 2nd feedback on PR