[FIX] packagemore_ecommerce_pricelists: Get min qty in all prlst itms
After latest IMP on website product list view, unexpected behavior
appeared: when having 2 or more pricelist items, setting the qty
above the min_quantity of a higher pricelist item did not update
the price to pay.

This is a fix to that unexpected behavior while keeping the feature
in product list view of showing the product price of the smallest
min_quantity among its pricelist items even when this smallest
min_quantity is greater than 1.

opw-3389883

[IMP] packagemore_ecommerce_pricelists: Set minimum qty and show price in website product list

When a product has more than one pricelist line, and all minimum quantities are
greater than 1, the website product list view would still show "Please contact
us for quote", which was not expected.
This commit takes the smallest minimum quantity defined for a product in a
pricelist, and displays the corresponding price in the website product list view.

Additionally, when you enter the form view of a product in the website, the
smallest minimum quantity in the pricelist is automatically set.