[FIX] packagemore_ecommerce_pricelists: Get min qty in all prlst itms
After latest IMP on website product list view, unexpected behavior
appeared: when having 2 or more pricelist items, setting the qty
above the min_quantity of a higher pricelist item did not update
the price to pay.

This is a fix to that unexpected behavior while keeping the feature
in product list view of showing the product price of the smallest
min_quantity among its pricelist items even when this smallest
min_quantity is greater than 1.

opw-3389883